Try to use `register_scheme` before @@schemes
Looks like URI#register_scheme was introduced in Ruby 3.1, but
backported to the standalone URI ruby gem so there's a non-zero
chance that an older Ruby version may end up with a newer URI library.
To be safe, we should try both versions by failing back to explicitly
using the class level variable if the new loader method fails.