Use local constants for SCHEME & TRIPLETS to fix spec tests
In Ruby 3.1.1, I started seeing, demonstrated by test:spec

    Error:
URI::#parse takes triplet "git@192.168.101.127:user/project.git"#test_0002_recognizes URI::SSH as a triplet:
    TypeError: type mismatch: String given

This indicated to me that something was going wrong with the monkey-patching of `RFC3986_Parser` with `TripletInterruptus`.

Following a terribly simple-minded approach, I renamed the constants `SCHEME` and `TRIPLET` to avoid overwriting existing constants. I could not specifically find what/where in Ruby 3.x this behavior changed, so I'm not 100% certain that this is the best fix.

It does, however, cause the specs to pass.